Remove comment on addgame and move styles to formstyled.tsx
